Floating context panel is shifted down whenever modal dialog opens #290
Updated transition to be applied to the sliding (`right`) only.
Updated position of the floating panel, when the upper container is blurred, since "a value other than none for the filter property results in the creation of a containing block for absolute and fixed positioned descendants". Thus, position of the floating panel must be adjusted.