Issue Details Dialog - increase top padding in the footer #871
Decreased space between comments.
Removed spacing on the right of the assignees list, making it similar to the items list.
Added word breaking, when comments overflow the comments section, e.g. when you comment with a long words without the whitespaces, like links.
Fixed different height of the dialog buttons on each tab.