Long search string breaks filter in Content Selector #3716
Updated error handling in `ContentSelectorQueryRequest`: previously, it was handled in the catch block of the `sendAndParse()` method, which resulted in skipping the `parseResponse()` call, so some flags were not set properly. Moved `catch` block to `send()` method, so the error is now properly handled and the data is passed further.
Fixed the Metadada object in responses, replacing it with the json-like interface instead of class, because it is not parsed.
Removed `ContentMetadata` and `IssueMetadata`, moving the code to a single `ResultMetadata` class.

(cherry picked from commit d8b4936d4169fd8d67a87f33a64eb01669ebb5ea)