Item Set, Occurrence View - tooltip for 'Add' button is not shown after saving a new added set #3437
Added removal of previous tooltip, if it's still shown.
Prevented delayed (via timeout) removal of the tooltip, if the tooltip target was changed.
The original issue appeared upon clicking on the button, which leads to DOM changes and showing a new element with a title under the current mouse position.